feat: Enhance subscription payment model and UI
- Added `userAiRequestedFoods` and `isActive` fields to `SubscriptionPayment` model.
- Updated JSON serialization for new fields in `subscription_payment.g.dart`.
- Refactored `RemoteApiBase` to streamline subscription payment JSON handling.
- Modified `UriBuilder` to switch to a local development server.
- Simplified `PaymentRepository` by removing unnecessary state management.
- Improved `PaymentDialog` UI with localized strings and dynamic content.
- Updated localization files to include new strings for payment dialog and privacy policy.
- Introduced `PrivacyDialog` component for displaying privacy information.
- Created `Transactions` screen to display user subscription transactions with error handling.
- Enhanced transaction display with detailed information and localized labels.

+++ packages/application/vo2max_calculator/lib/src/vo2max_calculator.dart
@@ -131,7 +131,7 @@ class _Vo2maxCalculatorCardState extends State<Vo2maxCalculatorCard> {
               Text(
                 context.l10n.vo2maxCalculatorVo2maxDistance(
                   15,
-                  maxDistanceInMeterInOneMinues * 15,
+                  maxDistanceInMeterInOneMinues.toInt() * 15,
                 ),
 
                 textAlign: TextAlign.justify,
@@ -139,21 +139,21 @@ class _Vo2maxCalculatorCardState extends State<Vo2maxCalculatorCard> {
               Text(
                 context.l10n.vo2maxCalculatorVo2maxDistance(
                   20,
-                  maxDistanceInMeterInOneMinues * 20,
+                  maxDistanceInMeterInOneMinues.toInt() * 20,
                 ),
                 textAlign: TextAlign.justify,
               ),
               Text(
                 context.l10n.vo2maxCalculatorVo2maxDistance(
                   25,
-                  maxDistanceInMeterInOneMinues * 25,
+                  maxDistanceInMeterInOneMinues.toInt() * 25,
                 ),
                 textAlign: TextAlign.justify,
               ),
               Text(
                 context.l10n.vo2maxCalculatorVo2maxDistance(
                   30,
-                  maxDistanceInMeterInOneMinues * 30,
+                  maxDistanceInMeterInOneMinues.toInt() * 30,
                 ),
                 textAlign: TextAlign.justify,
               ),
